requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.14
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
astroid==3.2.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==3.1.0
cmake==3.30.0
confection==0.1.5
contourpy==1.2.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
datasets==2.20.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@5aeb252b194b93dc2879b4ac34bc51a31b5aee13
diffusers==0.28.0
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.8.0
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fairscale==0.4.13
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.0
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
imageio==2.35.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
iopath==0.1.9
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
lightning-utilities==0.11.8
lit==18.1.8
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
peft==0.11.1
Pillow==10.2.0
platformdirs==4.2.2
portalocker==2.10.1
preshed==3.0.9
protobuf==3.21.6
psutil==6.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycocotools==2.0.8
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.2
pydub==0.25.1
Pygments==2.18.0
pylance==0.14.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytorch-msssim==1.0.0
pytorch_ssim==0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.12
ruff==0.5.2
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.1
semantic-version==2.10.0
sentry-sdk==2.13.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.40.0
sympy==1.13.0
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
thinc==8.2.5
thop==0.1.1.post2209072238
tifffile==2024.8.10
timm==1.0.11
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torchmetrics==1.4.3
transformers==4.41.1
triton==2.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wandb==0.17.7
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==11.0.3
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
